requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
pandas==2.1.3
python-dotenv==1.0.0
lxml==4.9.3
unidecode==1.3.7
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
webdriver-manager==4.0.1